backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==3.2.16
django-cors-headers==3.13.0
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==6.0.0
flake8-isort==6.0.0
idna==3.10
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
oauthlib==3.2.2
Pillow==9.0.0
pycodestyle==2.10.0
pycparser==2.22
pyflakes==3.0.1
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
